Now duration of podcast is sent in seconds as a double

Approved-by: Juan José Alfocea Marin

+++ app/src/main/kotlin/io/square1/limor/scenes/main/fragments/podcast/PodcastDetailsFragment.kt
@@ -1793,7 +1793,7 @@ class PodcastDetailsFragment : BaseFragment() {
         // duration
 //        uiPodcast?.audio?.duration?.let {
         uiPodcast?.audio?.total_length?.let {
-            tvPodcastTime?.text = CommonsKt.calculateDurationMinutesAndSeconds(it.toLong())
+            tvPodcastTime?.text = CommonsKt.calculateDurationMinutesAndSeconds(it)
         }
 
         // recasts
